+++ source/camera_module/camera.py
@@ -1,0 +1,40 @@
+
+import time
+import os
+from utils import get_project_root
+
+test_dir = str(get_project_root())+'/source/data_Set/data/'
+
+class Camera:
+    def __init__(self):
+        print("Init camera")
+
+        self.resolution = (2592, 1944)
+        self.rotation = 180
+        self.filesCount = self.capture(test_dir)
+
+    def set_camera(self, resolution, rotation):
+        print("Set up camera")
+        self.resolution = resolution
+        self.rotation = rotation
+        return resolution, rotation
+
+    def start_preview(self):
+        print("Start preview")
+
+    def stop_preview(self):
+        print("Stop preview")
+
+    def capture(self, file_path):
+        count = 0
+        for base, dirs, files in os.walk(file_path):
+            for Files in files:
+                print("Make photo frame%04d.jpg" % count)
+                count += 1
+
+        print("count = ", count)
+        return count
+
+
+    def sleep(self, sec):
+        time.sleep(sec)
